requirements.txt
anyio==3.6.1
asgiref==3.5.2
click==8.1.3
Django==4.0.5
fastapi==0.78.0
h11==0.13.0
httptools==0.4.0
idna==3.3
image==1.5.33
Pillow==9.1.1
pydantic==1.9.1
python-dotenv==0.20.0
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
starlette==0.19.1
typing_extensions==4.2.0
uvicorn==0.18.1
uvloop==0.16.0
watchfiles==0.15.0
websockets==10.3
